Move search engine XML to SearchEngineManager.
